+++ R/board_url.R
@@ -14,6 +14,8 @@
 #'   use the last cached version? Defaults to `is_interactive()` so you'll
 #'   be robust to poor internet connectivity when exploring interactively,
 #'   but you'll get clear errors when the code is deployed.
+#' @param headers Named character vector for additional HTTP headers (such as for
+#'   authentication). See [connect_auth_headers()] for Posit Connect support.
 #' @family boards
 #' @inheritParams new_board
 #' @details
@@ -42,8 +44,25 @@
 #' Using a [manifest file][write_board_manifest()] can be useful because you
 #' can serve a board of pins and allow collaborators to access the board
 #' straight from a URL, without worrying about board-level storage details.
-#'
 #' Some examples are provided in `vignette("using-board-url")`.
+#'
+#' # Authentication for `board_url()`
+#'
+#' The `headers` argument allows you to pass authentication details or other
+#' HTTP headers to the board, such as for a Posit Connect vanity URL that is
+#' not public (see [board_connect_url()]) or a private GitHub repo.
+#'
+#' ```r
+#' gh_pat_auth <- c(
+#'   Authorization = paste("token", "github_pat_XXXX")
+#' )
+#' board <- board_url(
+#'   "https://raw.githubusercontent.com/username/repo/main/path/to/pins",
+#'   headers = gh_pat_auth
+#' )
+#'
+#' board %>% pin_list()
+#' ```
 #'
 #' @export
 #' @examples
@@ -69,15 +88,18 @@
 #'
 board_url <- function(urls,
                       cache = NULL,
-                      use_cache_on_failure = is_interactive()) {
+                      use_cache_on_failure = is_interactive(),
+                      headers = NULL) {
 
+  check_headers(headers)
   url_format <- get_url_format(urls)
   if (url_format == "pins_yaml") {
-    manifest <- get_manifest(urls)
+    manifest <- get_manifest(urls, headers)
     board <- board_url(
       manifest,
       cache = cache,
-      use_cache_on_failure = use_cache_on_failure
+      use_cache_on_failure = use_cache_on_failure,
+      headers = headers
     )
     return(board)
   }
@@ -92,7 +114,8 @@ board_url <- function(urls,
     urls = urls,
     cache = cache,
     versioned = versioned,
-    use_cache_on_failure = use_cache_on_failure
+    use_cache_on_failure = use_cache_on_failure,
+    headers = headers
   )
 }
 
@@ -136,7 +159,8 @@ pin_meta.pins_board_url <- function(board, name, version = NULL, ...) {
       url = paste0(url, "data.txt"),
       path_dir = cache_dir,
       path_file = "data.txt",
-      use_cache_on_failure = board$use_cache_on_failure
+      use_cache_on_failure = board$use_cache_on_failure,
+      headers = board$headers
     )
     meta <- read_meta(cache_dir)
     local_meta(
@@ -187,7 +211,8 @@ pin_fetch.pins_board_url <- function(board, name, version = NULL, ...) {
       url = url,
       path_dir = meta$local$dir,
       path_file = file,
-      use_cache_on_failure = board$use_cache_on_failure
+      use_cache_on_failure = board$use_cache_on_failure,
+      headers = board$headers
     )
   })
 
@@ -240,7 +265,7 @@ get_url_format <- function(urls) {
   }
 }
 
-get_manifest <- function(url, call = rlang::caller_env()) {
+get_manifest <- function(url, headers, call = rlang::caller_env()) {
   # if ends with "/", look for manifest
   if (grepl("/$", url)) {
     url <- paste0(url, manifest_pin_yaml_filename)
@@ -249,7 +274,7 @@ get_manifest <- function(url, call = rlang::caller_env()) {
   # if request fails or returns with error code
   tryCatch(
     {
-      resp <- httr::GET(url)
+      resp <- httr::GET(url, httr::add_headers(headers))
       httr::stop_for_status(resp)
     },
     error = function(e) {
@@ -293,6 +318,7 @@ get_manifest <- function(url, call = rlang::caller_env()) {
 
 http_download <- function(url, path_dir, path_file, ...,
                           use_cache_on_failure = FALSE,
+                          headers = NULL,
                           on_failure = NULL) {
   cache_path <- download_cache_path(path_dir)
   cache <- read_cache(cache_path)[[url]]
@@ -303,12 +329,11 @@ http_download <- function(url, path_dir, path_file, ...,
       return(cache$path)
     }
 
-    headers <- httr::add_headers(
+    headers <- c(
+      headers,
       `If-Modified-Since` = http_date(cache$modified),
       `If-None-Match` = cache$etag
     )
-  } else {
-    headers <- NULL
   }
 
   path <- fs::path(path_dir, path_file)
@@ -317,7 +342,7 @@ http_download <- function(url, path_dir, path_file, ...,
   write_out <- httr::write_disk(tmp_path)
 
   req <- tryCatch(
-    httr::GET(url, headers, ..., write_out),
+    httr::GET(url, httr::add_headers(headers), ..., write_out),
     error = function(e) {
       if (!is.null(cache) && use_cache_on_failure) {
         NULL
@@ -402,4 +427,11 @@ http_date <- function(x = Sys.time(), tz = "UTC") {
 
   withr::local_locale(LC_TIME = "C")
   strftime(.POSIXct(x), "%a, %d %b %Y %H:%M:%S", tz = tz, usetz = TRUE)
+}
+
+check_headers <- function(x, arg = caller_arg(x), call = caller_env()) {
+  if (!is.null(x) && (!is_character(x) || !is_named(x))) {
+    stop_input_type(x, "a named character vector", allow_null = TRUE, arg = arg, call = call)
+  }
+
 }